Resync with local storage changes when used in multiple windows.
Previously if a group was added, removed, or changed in a different
window the group UI would not update leaving the state out of sync with
the local storage.

This patch fixes that.

Along with that change comes minimising the amount of local storage
save calls emitted to prevent over-saving when unnecessary.